--- Scrabble/nbproject/private/private.properties
@@ -1,2 +1,0 @@
-compile.on.save=true
-user.properties.file=C:\\Users\\Daniel\\AppData\\Roaming\\NetBeans\\8.2\\build.properties
